coord: double -> int

+++ cw-main/src/main/java/com/cwsni/world/model/Point.java
@@ -2,31 +2,31 @@ package com.cwsni.world.model;
 
 public class Point {
 	
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	public Point() {
 		this(0, 0);
 	}
 	
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 	
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 	
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
